# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API server"
        }
    ],
    "paths": {
        "/rating": {
            "post": {
                "tags": [
                    "Notation"
                ],
                "summary": "Noter un voyage",
                "operationId": "abdaadda14312676e1da4b96afd9f227",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_id",
                                    "rating"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Note entre 1 et 5",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notation ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notation ajoutée avec succès."
                                        },
                                        "rating": {
                                            "$ref": "#/components/schemas/Rating"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/rating/{trip_id}": {
            "get": {
                "tags": [
                    "Notation"
                ],
                "summary": "Obtenir les notations pour un voyage",
                "operationId": "98208b1140f5a1939e4aac0be2442326",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notations récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rating"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voyage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trip": {
            "get": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Obtenir la liste des voyages",
                "operationId": "cadf5c68469d51c35eaca2a49b49acb0",
                "responses": {
                    "200": {
                        "description": "Liste des voyages récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Voyage"
                ],
                "summary": "Créer un nouveau voyage",
                "operationId": "881dc7b53d5d143c7d8440a15d4c06ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starting_point",
                                    "ending_point",
                                    "starting_at",
                                    "available_places",
                                    "price",
                                    "user_id"
                                ],
                                "properties": {
                                    "starting_point": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ending_point": {
                                        "type": "string",
                                        "example": "Londres"
                                    },
                                    "starting_at": {
                                        "type": "string",
                                        "example": "2024-08-01T10:00:00Z"
                                    },
                                    "available_places": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voyage créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trip"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Rechercher des trajets",
                "operationId": "9ffcbfef8d645c0f7505d3a50bf2b8af",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Point de départ du trajet",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "Point d'arrivée du trajet",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date du trajet",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des trajets récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trip"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun trajet trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Obtenir la liste des utilisateurs",
                "operationId": "23679253d16f34d2972249751e3631dd",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lastname",
                                    "firstname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "lastname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Obtenir un utilisateur par ID",
                "operationId": "1ec60d4ee968175164a94c21f897f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Mettre à jour un utilisateur",
                "operationId": "786a794071cb7098f55a4a4911e7d3f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lastname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "trip_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "avatar.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Utilisateur"
                ],
                "summary": "Supprimer un utilisateur",
                "operationId": "a56b12a3694a926699892056cdcf8a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Rating": {
                "required": [
                    "user_id",
                    "trip_id",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "trip_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "description": "Note entre 1 et 5",
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Trip": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "starting_point": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "ending_point": {
                        "type": "string",
                        "example": "Lyon"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-29T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-29T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "trips_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatar.png"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-23T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-23T12:34:56Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}